requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-docx==1.1.2
beautifulsoup4==4.12.3
httpx==0.27.0
anthropic==0.39.0
python-multipart==0.0.12
pymupdf==1.24.0
